refactor: unify plant list into shared component
